# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ScholarAgent API
  description: |
    ScholarAgent - 学术智能体 API

    一个基于 AI 的学术文献检索与分析智能体系统，支持多数据源检索、智能查询规划、
    证据验证和文献推荐。

    ## 主要功能

    - **智能查询规划**: 使用 LLM 将复杂学术问题分解为多个子查询
    - **多数据源检索**: 集成 Semantic Scholar、OpenAlex、CrossRef、arXiv
    - **证据验证**: 从论文全文中提取支持/反驳证据
    - **文献推荐**: 基于用户反馈的个性化推荐
  version: 1.0.0
  contact:
    name: ScholarAgent Team
    email: team@scholar-agent.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
  - url: https://api.scholar-agent.dev
    description: 生产服务器
...
paths:
...
  /api/v1/search:
    post:
      operationId: createSearch
      summary: 创建复杂查询搜索
      description: |
        接收用户的自然语言查询，使用 LLM 进行查询规划，然后并行检索多个学术数据源。
        返回搜索运行 ID，可通过 SSE 获取实时进度。
      tags: [search]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            examples:
              basic:
                summary: 基础搜索
                value:
                  query: "transformer model for protein structure prediction"
                  max_results: 50
                  sources: ["semantic_scholar", "openalex"]
              advanced:
                summary: 高级搜索（带时间范围）
                value:
                  query: "deep learning in medical imaging, focus on CT scans"
                  max_results: 100
                  sources: ["semantic_scholar", "crossref", "arxiv"]
                  date_from: "2020-01-01"
                  date_to: "2024-12-31"
                  min_citations: 10
      responses:
        '202':
          description: 搜索任务已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/search/{run_id}:
    get:
      operationId: getSearchRun
      summary: 获取搜索运行详情
      description: 获取指定搜索运行的状态、进度和结果
      tags: [search]
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: 搜索运行详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchRunDetail'
        '404':
          $ref: '#/components/responses/NotFound'

  # =============================================================================
  # 论文接口
  # =============================================================================
...
components:
...
  parameters:
    RunId:
      name: run_id
      in: path
      required: true
      description: 搜索运行 ID
      schema:
        type: string
        format: uuid
...
  schemas:
...
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          description: 用户的自然语言查询
          minLength: 1
          maxLength: 2000
          example: "transformer model for protein structure prediction"
        max_results:
          type: integer
          description: 最大返回结果数
          default: 50
          minimum: 1
          maximum: 500
        sources:
          type: array
          description: 指定数据源
          items:
            $ref: '#/components/schemas/DataSource'
          default: ["semantic_scholar", "openalex"]
        date_from:
          type: string
          format: date
          description: 起始日期
        date_to:
          type: string
          format: date
          description: 结束日期
        min_citations:
          type: integer
          description: 最小引用数
          minimum: 0
        open_access_only:
          type: boolean
          description: 仅开放获取论文
          default: false
...
    SearchResponse:
      type: object
      properties:
        run_id:
          type: string
          format: uuid
          description: 搜索运行 ID
        status:
          $ref: '#/components/schemas/SearchStatus'
        query_plan:
          $ref: '#/components/schemas/QueryPlan'
        message:
          type: string
          description: 状态消息
    SearchRunDetail:
      type: object
      properties:
        run_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/SearchStatus'
        original_query:
          type: string
        query_plan:
          $ref: '#/components/schemas/QueryPlan'
        progress:
          $ref: '#/components/schemas/SearchProgress'
        results:
          type: array
          items:
            $ref: '#/components/schemas/Paper'
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    QueryPlan:
      type: object
      properties:
        sub_queries:
          type: array
          items:
            type: object
            properties:
              query:
                type: string
              source:
                $ref: '#/components/schemas/DataSource'
              rationale:
                type: string
          description: LLM 生成的子查询列表
        strategy:
          type: string
          description: 检索策略说明
    SearchProgress:
      type: object
      properties:
        total_sources:
          type: integer
        completed_sources:
          type: integer
        total_papers:
          type: integer
        deduplicated_papers:
          type: integer
        current_phase:
          type: string
          enum: [planning, searching, deduplicating, ranking, completed, failed]
    Paper:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        abstract:
          type: string
        year:
          type: integer
        venue:
          type: string
          nullable: true
        citation_count:
          type: integer
        doi:
          type: string
          nullable: true
        url:
          type: string
          format: uri
        pdf_url:
          type: string
          format: uri
          nullable: true
        source:
          $ref: '#/components/schemas/DataSource'
        relevance_score:
          type: number
          minimum: 0
          maximum: 1
        is_open_access:
          type: boolean
...
    DataSource:
      type: string
      enum:
        - semantic_scholar
        - openalex
        - crossref
        - arxiv
      description: 学术数据源
    SearchStatus:
      type: string
      enum:
        - pending
        - running
        - completed
        - failed
      description: 搜索状态

    # ----- 错误响应 -----
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: 错误详情
        code:
          type: string
          description: 错误代码
        timestamp:
          type: string
          format: date-time

  # =============================================================================
  # 响应模板
  # =============================================================================
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Invalid request parameters"
            code: "BAD_REQUEST"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Resource not found"
            code: "NOT_FOUND"
    ValidationError:
      description: 数据验证错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Validation error"
            code: "VALIDATION_ERROR"
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Internal server error"
            code: "INTERNAL_ERROR"